+++ package.json
@@ -16,7 +16,7 @@
     "lint-staged": "lint-staged --config lint-staged.cjs"
   },
   "dependencies": {
-    "@auth0/auth0-react": "^1.10.2",
+    "@auth0/auth0-react": "1.11.0",
     "@reduxjs/toolkit": "^1.8.5",
     "axios": "^0.27.2",
     "react": "^18.2.0",
@@ -26,34 +26,34 @@
     "vite-plugin-css-injected-by-js": "^2.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.19.0",
-    "@storybook/addon-actions": "^6.5.10",
-    "@storybook/addon-docs": "^6.5.10",
-    "@storybook/addon-essentials": "^6.5.10",
-    "@storybook/addon-interactions": "^6.5.10",
-    "@storybook/addon-links": "^6.5.10",
-    "@storybook/addon-viewport": "^6.5.10",
-    "@storybook/addons": "^6.5.10",
+    "@babel/core": "7.19.1",
+    "@storybook/addon-actions": "6.5.12",
+    "@storybook/addon-docs": "6.5.12",
+    "@storybook/addon-essentials": "6.5.12",
+    "@storybook/addon-interactions": "6.5.12",
+    "@storybook/addon-links": "6.5.12",
+    "@storybook/addon-viewport": "6.5.12",
+    "@storybook/addons": "6.5.12",
     "@storybook/builder-vite": "^0.2.2",
     "@storybook/jest": "^0.0.10",
-    "@storybook/react": "^6.5.10",
+    "@storybook/react": "6.5.12",
     "@storybook/testing-library": "^0.0.13",
-    "@storybook/theming": "^6.5.10",
+    "@storybook/theming": "6.5.12",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/react-hooks": "^8.0.1",
     "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^29.0.1",
-    "@types/node": "^18.7.16",
-    "@types/react": "^18.0.17",
+    "@types/jest": "29.0.2",
+    "@types/node": "18.7.18",
+    "@types/react": "18.0.20",
     "@types/react-dom": "^18.0.6",
     "@types/redux-persist": "^4.3.1",
-    "@typescript-eslint/eslint-plugin": "^5.36.2",
-    "@typescript-eslint/parser": "^5.36.2",
+    "@typescript-eslint/eslint-plugin": "5.37.0",
+    "@typescript-eslint/parser": "5.37.0",
     "@vitejs/plugin-react": "^2.1.0",
-    "autoprefixer": "^10.4.8",
+    "autoprefixer": "10.4.11",
     "babel-loader": "^8.2.5",
-    "eslint": "^8.23.0",
+    "eslint": "8.23.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-jest": "^27.0.4",
     "eslint-plugin-prettier": "^4.2.1",
@@ -73,10 +73,10 @@
     "stylelint-config-recess-order": "^3.0.0",
     "stylelint-config-standard": "^28.0.0",
     "tailwindcss": "^3.1.8",
-    "ts-jest": "^29.0.0",
+    "ts-jest": "29.0.1",
     "tsc-files": "^1.1.3",
     "typescript": "^4.8.3",
-    "vite": "^3.1.0",
+    "vite": "3.1.1",
     "vite-plugin-windicss": "^1.8.8",
     "vitest": "^0.23.2"
   }
