+++ src/components/Facilities/Facilities.tsx
@@ -1,0 +1,9 @@
+import { Button } from 'src/components/core'
+
+export const Facilities = () => {
+  return (
+    <div>
+      <Button></Button>
+    </div>
+  )
+}
