+++ src/components/core/index.tsx
@@ -1,1 +1,1 @@
-export { Button } from './Button'
+export { Button } from './Button/Button'
