+++ src/stories/Button.stories.tsx
@@ -5,21 +5,15 @@ import { Button } from './Button'
 import { expect } from '@storybook/jest'
 import { within, userEvent } from '@storybook/testing-library'
 
-// More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
   title: 'Example/Button',
   component: Button,
-  // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
-  argTypes: {
-    backgroundColor: { control: 'color' }
-  }
+  argTypes: {}
 } as ComponentMeta<typeof Button>
 
-// More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
 const Template: ComponentStory<typeof Button> = (args) => <Button {...args} />
 
 export const Primary = Template.bind({})
-// More on args: https://storybook.js.org/docs/react/writing-stories/args
 Primary.args = {
   primary: true,
   label: 'Button',
@@ -38,6 +32,12 @@ Large.args = {
 }
 
 export const Small = Template.bind({})
+Small.args = {
+  size: 'small',
+  label: 'Button'
+}
+
+export const Test = Template.bind({})
 Small.args = {
   size: 'small',
   label: 'Button'
