+++ src/stories/Button.tsx
@@ -1,5 +1,3 @@
-import './button.css'
-
 interface ButtonProps {
   /**
    * Is this the principal call to action on the page?
