+++ src/stories/Page.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 
 import { Header } from './Header'
-import './page.css'
 
 type User = {
   name: string
