+++ tsconfig.json
@@ -20,7 +20,6 @@
       "#rtkstore/*": ["src/app/store/*"],
       "#rtkhooks/*": ["src/app/hooks/*"],
       "#rtkfeatures/*": ["src/app/features/*"],
-      "#core/*": ["src/components/core/*"],
       "#utils/*": ["src/utils/*"]
     }
   },
